Adds flyout-kmaps-typeahead jQuery plugin and code clean
Modified the flyout-kmaps-typeahead.js jQuery plugin to be more agnostic
before it was under a .erb so it can pull variables from rails that has
been abstracted so those varialbes are pulled in teh view and the plugin
just normally. Previously the plugin was located in the places_engine
were it was first designed but it was moved to kmaps so it can be used
by any other new app.