Improved fancyTree plugins and Typeahead
Some modifications were done to the fancyTree plugins to better server
the uses we have currently it'll display in a nicer diagramation and
fixed an issue that required to double click the expand icon to activate
the lazy loading.

At the same time modified the typeahead plugin and due to those
modifications moved to our own app/assets/javascripts/kmaps_engine so it
explicitly shows that it is different than the vendor version. It was
modified to support not only Places & Subjects now it can support
Sources too so it will be used in the editorial interface.