Improve kmapstree (fly-out and related) for perspectives
Improved the code to support perspectives not just for Places but for
subjects too, The Subjects index still needs to be fixed to provide a
standarized id format, for places the id has the form: domain-id while
for subjects is just the number id. This should be standarised to work
with new apps.

At the same time fixed a small problem on the editorial interface in
citations, it used to display two buttons with "UPDATE" one shoudl say
CANCEL.