Add Solr utils JavaScript library
This new library works as a repository of the queries we would use in
all of our JavaScript calls, to make it easier to maintain and track.

This branch contains the following changes:

Updated Relations tree & Collapsible List Plugin
Rename the jQuery plugins to handle Kmaps Relations Tree and Collapsible
list to match the name of their individual github repos for consistency.

Updated KmapsPopup to use solr-utils library
Updated the code from jquery.kmaps-popup.js to now use the solr-utils.js
library this to keep all the solr queries in one place and simplify the
code.

Bug fix kmapstree and solr-utils
Fix kmapstree order alphabetically
Fix solr-utils to take not assume title of a label is in span, it now
expects just the name of the feature.

Add filter by app on the flyout typeahead
Before the flyout used to show all of the results no matter which app
the feature belongs to, now itis filtered by the current tree.