Fixed feature_name_header to use .s for Tibetan
When displaying Tibetan names the header was not showing the correct markup, this changes fixes this problem, but because of this change in functionality, we need to update the code where this method was called to take into account that it will have the markup tags. This change affects places and subjects code too.

Improve fancytree css