Refactor the use of flyout search results popover
Standardized the popups being used on the flyout. This change affected
how the popovers are displayed now a orientation property is used to be
able to display the popover to the side of the content correctly on the
flyout.

Add caption support for kmaps-popup

Currently it only displays caption in english, need to take into account
language.

The captions are obtained by Solr query via the solr-utils library.

Add support for language to popovers

Kmaps-popup now takes into consideration a specified language, you can
define which language to present for the caption.

Add Kmaps relations full tree and feature relations defaults

+ Add support for full kmapsRelationsTree
Removed old jquery.kmapstree.js, it was now replaced by the new
functionality added to kmaps_relations_tree. This change required
refactors on solr-utils library.

+ Replace the flyout tree with the new implementation for
kmapsRelationsTree.

+ Updated tree used on select ancestor when creating a new Feature.

+ Add feature relations defaults when creating a new feature. This
change required modifications on the session manager and application
settings to provide the methods to access this defaults.

Trimmed down call for optimal calculations and transmit sizes for Solr queries

Updated version in preparation for merge to master