Fix incompatibility with Safari for KmapsRelationsTree
There was an issue with reduceRight and the accumulator being sent as a
promise.

Removed the wait from the scrollToActive, the previous implementation
didn't take into account that the whole path should be expanded, now the
property is correctly set in the ancestors to show it is expanded.

Updated engine version.