Refactored progress bar and added option to sleep if its working hours
Moved the code that was being used to prevent the script from run during
working hours to progress bar, it might be better in another area but
because they are currently used together the functionality was bundled
together.

Fix length verification for Captions and Summaries

There was a bug when the characters &amp; were not being stripped with
strip_tags. It was fixed and should be working now.

This was reported on: MANU-5018

Update version in preparation for merge.