Fix bug on Relations Tree and Added simple typeahead update filter
There is a bug on Relations tree when a mark is not set for predicates,
 which seldom happens but can trigger an error.

Add new functionality to Kmaps simple typeahead to support updating the
additional filters setting.

Added mark all nodes, add support for Marking root nodes

Previously there was no way to mark all nodes in a tree, this allows me
to mark them as customAction and the redirect when activating a node is
not triggering.

Now solr-utils is able to mark root nodes.

Updated engine version