Adds Mandala Texts Integration to Kmaps as Essays for each Feature
Previously we had Descriptions presented to the User as Essays, that was
supposed to be migrated to Mandala Texts. Now we are creating a new
Essay modle that will be linked to Mandala Texts through
ShantiIntegration. This essays are going to be used to display a more
descriptive text for each Feature.

The Essay is tied up to the current Language. So a Feature could have
many Essays with different languages. When the Feature is displayed we
are going to show the Essay for the language currently selected by the
user in view mode.

Improves Mandala Texts association editorial interface

Before we were using a typeahead input text to search the whole Mandala
Texts index. Now we are filtering by the texts already associated with
the current feature. We display only the associated texts to the user
and if none is found a message is displayed to the user to notify her
that the text needs to be associated beforehand.

Updates Mandala texts integration editorial interface JavaScript

Added new logic to the frontend to provide better defaults on language
selection if the Mandla Text already contains a Language defined. We are
basing the Language options on the Rails applications based on the
available locales.

Adds homepage texts association to Solr index

Now we are adding the hompage text association to Solr using a dynamic
field with the following form:

`homepage_text_{subject-id}_s`

And we are now keeping the mandala mapping on the Language model.